requirements.txt
# Python依赖

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate>=4.0.0
Flask-Cors==6.0.1
Flask-Bcrypt>=1.0.1
PyJWT>=2.7.0
python-dotenv>=1.0.0
Werkzeug==2.3.7
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.19
python-dateutil==2.8.2
numpy>=1.26.0
pandas>=2.2.0